server/requirements.txt
Django==2.0.5
django-extensions==2.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
psycopg2-binary==2.7.5
python-json-logger==0.1.9